+++ packages/redux_persist_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.4.3] - 2018-03-12
+
+* Fix load bug (loaded JSON instead of string)
+
 ## [0.4.2] - 2018-03-10
 
 * Breaking: Change PersistorGate `child` to `builder`.